Replace template content with draft spec
The initial commit is a clone of the [GitHub template](https://github.com/peter-evans/swagger-github-pages) we used to create this repo.

As a starter for ten we want to update the demo content with the complete draft version of our current spec from [sroc-service-team](https://github.com/DEFRA/sroc-service-team). We also update the README and licence stuff to ensure the repo is compliant with Defra standards.