Replace template content with draft spec (#1)
The initial commit is a clone of the [GitHub template](https://github.com/peter-evans/swagger-github-pages) we used to create this repo.

As a starter for ten, we want to update the demo content with the complete draft version of our current spec from [sroc-service-team](https://github.com/DEFRA/sroc-service-team). We also update the README and licence stuff to ensure the repo is compliant with Defra standards.

** Notes

We rename the root API spec file from `swagger.yaml` to `openapi.yml`. OpenAPI is the name for the standard going forward and also currently a synonym for version 3 (which we're using). Then updated the content by copying what we have in our draft.yml version at the moment.

Not much content in the README at the moment. But it at least brings us in line with the Defra standard. We'll add more content once we have figured out how best to set up this repo.