Rename root openapi.yml to draft.yml
Should have clocked this when initially updating the project from the template content.

`openapi.yml` is the name of the file we use in our actual spec. This a root file that contains lots of JSON references to the actual content of the spec.

When we then generate the version we publish we call this `draft.yml`. That version is wha the template expects as the root file and that's why we're better off renaming it to that. At the very least we remain consistent with what we were doing until we know better.