Add previous versions
In our previous worfklow whenever a new version of the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) was created a matching OpenAPI spec would also be created.

We store them here so we have a record of what functionality we said was available in a particular version of the API.

This change adds copies of the those files to this repo. The one different is we are no longer treating `draft.yml` as a version as it has become the projects root file.