Add previous versions (#4)
In our previous workflow whenever a new version of the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) was created a matching OpenAPI spec would also be created.

We store them here so we have a record of what functionality we said was available in a particular version of the API.

This change adds copies of those files to this repo. The one difference is we are no longer treating `draft.yml` as a version as it has become the project's root file.