Add npm package files
This with the included package script means anyone working on the project can quickly update the root `draft.yml` by simply running `npm start` from the root of the project.

We keep the package-lock.json to ensure everyone is using the same version of swagger-cli. Finally, we update the .gitignore to stop `node_modules` getting committed.