Add spec generator to the project (#5)
In the [sroc-service-team](https://github.com/DEFRA/sroc-service-team) version of producing the OpenAPI spec, we relied on a VSode extension called [openapi-designer](https://marketplace.visualstudio.com/items?itemName=philosowaffle.openapi-designer) to compile the de-referenced spec file.

It also handled previewing the file. The downsides were that previewing could be flaky and sometimes required a restart. It also compiled to JSON which meant we then had to use [JSON2YAML](https://www.json2yaml.com/) to convert it to YAML.

We since hit upon the NPM package [Swagger/OpenAPI CLI](https://www.npmjs.com/package/swagger-cli) which can handle the conversion without an interim step. We have also found [Swagger Viewer](https://marketplace.visualstudio.com/items?itemName=Arjun.swagger-viewer) does a better job of previewing the in-progress file.

So, this change adds the swagger-cli to the project so we have a set way of generating `draft.yml` in a consistent format. What developers choose to use to preview the file can be left to them as we are no longer reliant on the VSCode extension.

We also update `draft.yml` using swagger-cli. The formatting has changed because it seems no 2 OpenAPI generators are the same!

Key differences seem to be it doesn't use YAML's `|` syntax to denote multiline text and feels the need to escape ' wherever it appears in a sentence.