Update title in UI and spec
We spotted when viewed through GitHub pages the title was always Swagger UI. So, we came intending to correct that and then realised we also need to drop SROC from the title in the spec (the API handles both SROC and PRESROC).

So, this change corrects the index.html page which will resolve the title shown in the browser. It also updates the openapi.yml and draft.yml with the same title.