Fix update swagger-ui workflow (#7)
The [template project](https://github.com/peter-evans/swagger-github-pages) we built this one from included a GitHub workflow to automatically check we're using the latest version of [swagger-ui](https://github.com/swagger-api/swagger-ui). If not it automatically handles creating a pull request with the update.

But we have only just spotted that it is expecting the root OpenAPI spec file to be called `swagger.yaml` and we opted to move away from that in an earlier commit. It also pulls down a new version of `index.html` and overwrites the one present. This means our change to the page title will also be lost.

So, this change updates the workflow to fix these issues before the next update PR is generated.