CMEA-232 Add v3 delete invoice endpoint
https://eaflood.atlassian.net/browse/CMEA-232

As part of supporting SROC within the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) we're uplifting all endpoints to `/v3`. Some actually need changes, others the functionality remains the same. For clients though, it will be less confusing if all endpoints in the SROC supported workflow are `/v3`.

This change adds a `DELETE /v3` bill run invoice endpoint which is exactly the same as `/v2`. Within the API both routes point to the same code.