CMEA-232 Add v3 delete invoice endpoint (#9)
https://eaflood.atlassian.net/browse/CMEA-232

As part of supporting SROC within the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) we're uplifting all endpoints to `/v3`. Some actually need changes, others the functionality remains the same. For clients though, it will be less confusing if all endpoints in the SROC supported workflow are `/v3`.

This change adds a `DELETE /v3` bill run invoice endpoint which is exactly the same as `/v2`. Within the API both routes point to the same code.

** Notes 

As part of this change, we also update the v2 operationId to be unique

As per the [spec](https://spec.openapis.org/oas/v3.0.2#operation-object)

> Unique string used to identify the operation. The id MUST be unique among all operations described in the API. The operationId value is case-sensitive. Tools and libraries MAY use the operationId to uniquely identify an operation, therefore, it is RECOMMENDED to follow common programming naming conventions.