Fix notes on viewing versions in the README
It has been pointed out that we have misunderstoodthe purpose of the **Explore** box in the Swagger UI (when viewing the documentation).

We thought it was for searching through the current document. Turns out it displays whatever spec is entered in the field. This is why it always appears with `draft.yml` in the field. If you change that to `versions/v0-15-0.yml`, for example, it will switch to displaying that spec instead.

So, this change removes the incorrect note about **Explore** not working. It also completely revamps our guidance on how to view previous versions to exploit these feature of the Swagger UI.