Update validation lists for `loss` and `supportedSourceName` (#14)
The loss field has a slightly different set of values for sroc and presroc (where sroc is missing the Very Low value). We therefore update the docs to reflect this.

We also fix an issue where some values for supportedSourceName have an en-dash instead of a hyphen. This is being corrected in the CM and the rules service so we update the docs to match.


* Change dashes to hyphens

Some values for `supportedSourceName` had incorrectly had the hyphens autocorrected to en-dashes in the source document. We are correcting this in the CM and the rules service to change them back to hyphens, so we correct it here as well.

* Create separate `loss` fields for sroc and presroc

The list of allowed values for `loss` differs slightly between presroc and sroc. We therefore create separate lists of values for the two rulesets.

* Compile `draft.yml`