Add missing rebilledTransactionId to docs
Back in [SROC Charging Module API PR #456](https://github.com/DEFRA/sroc-charging-module-api/pull/456) we added the ability to link a rebilled tranasction to its originating transaction record.

As part of this we also included the `rebilledTransactionId` in the `GET /invoices/{id}` response but omitted adding it to the docs. It's been missing all this time but we've only just spotted that.

This change fixes the issue in the docs.