Rename draft.yml to draft.yaml
This is the first auto-update of the underlying project we depend on to host our OpenAPI docs on GitHub. Clearly, the folks on [swagger-github-pages](https://github.com/peter-evans/swagger-github-pages) believe in using `*.yaml` for the extension which is why the `index.html` was updated as part of the update.

Ideally, we want these to be as hands-off as possible. That's not going to happen if we keep having to revert the change to `index.html`. So, this change accepts the inevitable and renames our file to match **swagger-github-pages** convention.