Update reference to document
This is based on what we had in the initialiser of the `index.html` before v4.11.1 moved it out. We're working on the basis that the relative position of `draft.yml` will still be from the `index.html` and not the initialiser file. But that one will be a case of suck it and see after this change gets merged.